rename: story 파일들 stories 디렉터리로 이동
